requirements.txt
landingai-ade
pathlib
python-dotenv
pathway
openai
pandas
pydantic
watchdog
